requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
ccimport==0.3.7
certifi==2024.8.30
cffi==1.15.1
# chamfer==2.0.0
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
cumm-cu113==0.2.9
cycler==0.11.0
Cython==0.29.32
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
easydict==1.10
# emd-ext==0.0.0
entrypoints==0.4
fastjsonschema==2.16.3
filelock==3.9.1
fire==0.5.0
Flask==2.2.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
funcy==2.0
future==0.18.3
gensim==4.2.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
grpcio==1.53.0
huggingface-hub==0.13.2
idna==3.4
imageio==2.27.0
importlib-metadata==5.2.0
importlib-resources==5.10.2
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.5
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lark==1.1.5
lightning-utilities==0.5.0
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mpi4py==3.1.4
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.11.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.21.6
nuscenes-devkit==1.1.9
oauthlib==3.2.2
open3d==0.16.0
opencv-python==4.7.0.68
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pccm==0.2.21
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plotly==5.13.1
plyfile==0.8.1
portalocker==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.2
pycocotools==2.0.6
pycparser==2.21
pyDeprecate==0.3.0
Pygments==2.14.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.3.8
pytz==2022.7
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.0
SharedArray==3.2.2
six==1.16.0
# sklearn==0.0.post5
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spconv-cu113==2.1.21
strictyaml==1.6.2
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.1.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.12
tinycss2==1.2.1
torch==1.11.0+cu113
torch-scatter==2.1.0
torchaudio==0.11.0+cu113
torchmetrics==0.5.0
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trimesh==3.21.3
typing_extensions==4.4.0
urllib3==1.26.15
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
widgetsnbextension==4.0.5
yacs==0.1.8
yarl==1.8.2
zipp==3.11.0
